+++ example/.prettierrc.js
@@ -1,6 +1,8 @@
 module.exports = {
-  bracketSpacing: false,
-  jsxBracketSameLine: true,
-  singleQuote: true,
-  trailingComma: 'all',
+  bracketSpacing: true,
+  jsxBracketSameLine: false,
+  singleQuote: false,
+  trailingComma: "all",
+  tabWidth: 2,
+  semi: true,
 };
