+++ example/lib/local-assets/apple.png
