+++ example/lib/local-assets/facebook.png
