+++ example/lib/local-assets/google.png
