+++ example/lib/screens/sign-up/SignUpScreen.tsx
